py/FastAPI-demo/requirements.txt
fastapi
pydantic
Pillow
